Extract search utility functions to autoloadable utility class.
