+++ www/includes/easyparliament/searchengine.php
@@ -615,234 +615,3 @@ class SEARCHENGINE {
 
 global $SEARCHENGINE;
 $SEARCHENGINE = null;
-
-function search_by_usage($search, $house = 0) {
-        $data = array();
-        $SEARCHENGINE = new SEARCHENGINE($search);
-        $data['pagetitle'] = $SEARCHENGINE->query_description_short();
-        $SEARCHENGINE = new SEARCHENGINE($search . ' groupby:speech');
-        $count = $SEARCHENGINE->run_count(0, 5000, 'date');
-        if ($count <= 0) {
-            $data['error'] = 'No results';
-
-            return $data;
-        }
-        $SEARCHENGINE->run_search(0, 5000, 'date');
-        $gids = $SEARCHENGINE->get_gids();
-        if (count($gids) <= 0) {
-            $data['error'] = 'No results';
-
-            return $data;
-        }
-        if (count($gids) == 5000)
-            $data['limit_reached'] = true;
-
-        # Fetch all the speakers of the results, count them up and get min/max date usage
-        $speaker_count = array();
-        $gids = join('","', $gids);
-        $db = new ParlDB;
-        $q = $db->query('SELECT gid,person_id,hdate FROM hansard WHERE gid IN ("' . $gids . '")');
-        for ($n=0; $n<$q->rows(); $n++) {
-            $gid = $q->field($n, 'gid');
-            $person_id = $q->field($n, 'person_id');
-            $hdate = $q->field($n, 'hdate');
-            if (!isset($speaker_count[$person_id])) {
-                $speaker_count[$person_id] = 0;
-                $maxdate[$person_id] = '1001-01-01';
-                $mindate[$person_id] = '9999-12-31';
-            }
-            $speaker_count[$person_id]++;
-            if ($hdate < $mindate[$person_id]) $mindate[$person_id] = $hdate;
-            if ($hdate > $maxdate[$person_id]) $maxdate[$person_id] = $hdate;
-        }
-
-        # Fetch details of all the speakers
-        $speakers = array();
-        $pids = array();
-        if (count($speaker_count)) {
-            $person_ids = join(',', array_keys($speaker_count));
-            $q = $db->query('SELECT member_id, member.person_id, title, given_name, family_name, lordofname,
-                                constituency, house, party,
-                                moffice_id, dept, position, from_date, to_date, left_house
-                            FROM member LEFT JOIN moffice ON member.person_id = moffice.person
-                                JOIN person_names pn ON member.person_id = pn.person_id AND pn.type="name" AND pn.start_date <= left_house AND left_house <= pn.end_date
-                            WHERE member.person_id IN (' . $person_ids . ')
-                            ' . ($house ? " AND house=$house" : '') . '
-                            ORDER BY left_house DESC');
-            for ($n=0; $n<$q->rows(); $n++) {
-                $mid = $q->field($n, 'member_id');
-                if (!isset($pids[$mid])) {
-                    $title = $q->field($n, 'title');
-                    $first = $q->field($n, 'given_name');
-                    $last = $q->field($n, 'family_name');
-                    $lordofname = $q->field($n, 'lordofname');
-                    $house = $q->field($n, 'house');
-                    $party = $q->field($n, 'party');
-                    $full_name = ucfirst(member_full_name($house, $title, $first, $last, $lordofname));
-                    $pid = $q->field($n, 'person_id');
-                    $pids[$mid] = $pid;
-                    $speakers[$pid]['house'] = $house;
-                    $speakers[$pid]['left'] = $q->field($n, 'left_house');
-                }
-                $dept = $q->field($n, 'dept');
-                $posn = $q->field($n, 'position');
-                $moffice_id = $q->field($n, 'moffice_id');
-                if ($dept && $q->field($n, 'to_date') == '9999-12-31')
-                    $speakers[$pid]['office'][$moffice_id] = prettify_office($posn, $dept);
-                if (!isset($speakers[$pid]['name'])) {
-                    $speakers[$pid]['name'] = $full_name . ($house==1?' MP':'');
-                }
-                if ( !isset($speakers[$pid]['party']) && $party ) {
-                    $speakers[$pid]['party'] = $party;
-                }
-            }
-        }
-        if (isset($speaker_count[0])) {
-            $speakers[0] = array('party'=>'', 'name'=>'Headings, procedural text, etc.', 'house'=>0, 'count'=>0);
-        }
-        $party_count = array();
-        $ok = 0;
-        foreach ($speakers as $pid => &$speaker) {
-            $speaker['count'] = $speaker_count[$pid];
-            $speaker['pmaxdate'] = $maxdate[$pid];
-            $speaker['pmindate'] = $mindate[$pid];
-            $ok = 1;
-            if (isset($speaker['party'])) {
-                if (!isset($party_count[$speaker['party']])) {
-                    $party_count[$speaker['party']] = 0;
-                }
-                $party_count[$speaker['party']] += $count;
-            }
-        }
-
-        function sort_by_count($a, $b) {
-            if ($a['count'] > $b['count']) return -1;
-            if ($a['count'] < $b['count']) return 1;
-            return 0;
-        }
-        uasort($speakers, 'sort_by_count');
-        arsort($party_count);
-        if (!$ok) {
-            $data['error'] = 'No results';
-            return $data;
-        }
-
-        $data['party_count'] = $party_count;
-        $data['speakers'] = $speakers;
-
-        return $data;
-}
-
-// Return query result from looking for MPs
-function search_member_db_lookup($searchstring, $current_only=false) {
-    if (!$searchstring) return false;
-    $searchwords = explode(' ', $searchstring, 3);
-    $params = array();
-    if (count($searchwords) == 1) {
-        $params[':like_0'] = '%' . $searchwords[0] . '%';
-        $where = "given_name LIKE :like_0 OR family_name LIKE :like_0 OR lordofname LIKE :like_0";
-    } elseif (count($searchwords) == 2) {
-        // We don't do anything special if there are more than two search words.
-        // And here we're assuming the user's put the names in the right order.
-        $params[':like_0'] = '%' . $searchwords[0] . '%';
-        $params[':like_1'] = '%' . $searchwords[1] . '%';
-        $params[':like_0_and_1'] = '%' . $searchwords[0] . ' '. $searchwords[1] . '%';
-        $params[':like_0_and_1_hyphen'] = '%' . $searchwords[0] . '-'. $searchwords[1] . '%';
-        $where = "(given_name LIKE :like_0 AND family_name LIKE :like_1)";
-        $where .= " OR (given_name LIKE :like_1 AND family_name LIKE :like_0)";
-        $where .= " OR (title LIKE :like_0 AND family_name LIKE :like_1)";
-        $where .= " OR given_name LIKE :like_0_and_1 OR given_name LIKE :like_0_and_1_hyphen";
-        $where .= " OR family_name LIKE :like_0_and_1 OR family_name LIKE :like_0_and_1_hyphen";
-        $where .= " OR lordofname LIKE :like_0_and_1";
-        if (strtolower($searchwords[0]) == 'nick') {
-            $where .= " OR (given_name LIKE '%nicholas%' AND family_name LIKE :like_1)";
-        }
-    } else {
-        $searchwords[2] = str_replace('of ', '', $searchwords[2]);
-        $params[':like_0'] = '%' . $searchwords[0] . '%';
-        $params[':like_1'] = '%' . $searchwords[1] . '%';
-        $params[':like_2'] = '%' . $searchwords[2] . '%';
-        $params[':like_0_and_1'] = '%' . $searchwords[0] . ' '. $searchwords[1] . '%';
-        $params[':like_1_and_2'] = '%' . $searchwords[1] . ' '. $searchwords[2] . '%';
-        $params[':like_1_and_2_hyphen'] = '%' . $searchwords[1] . '-'. $searchwords[2] . '%';
-        $where = "(given_name LIKE :like_0_and_1 AND family_name LIKE :like_2)";
-        $where .= " OR (given_name LIKE :like_0 AND family_name LIKE :like_1_and_2)";
-        $where .= " OR (given_name LIKE :like_0 AND family_name LIKE :like_1_and_2_hyphen)";
-        $where .= " OR (title LIKE :like_0 AND family_name LIKE :like_1_and_2)";
-        $where .= " OR (title LIKE :like_0 AND family_name LIKE :like_1_and_2_hyphen)";
-        $where .= " OR (title LIKE :like_0 AND given_name LIKE :like_1 AND family_name LIKE :like_2)";
-        $where .= " OR (title LIKE :like_0 AND family_name LIKE :like_1 AND lordofname LIKE :like_2)";
-    }
-
-    $db = new ParlDB;
-    $q = $db->query("SELECT person_id FROM person_names WHERE type='name' AND ($where)", $params);
-    return $q;
-}
-
-function search_member_db_lookup_with_names($searchstring, $current_only=false) {
-    $q = search_member_db_lookup($searchstring, $current_only);
-
-    if ( !$q->rows ) {
-        return $q;
-    }
-
-    $person_ids = array();
-    for ($i=0; $i<$q->rows(); ++$i) {
-        $pid = $q->field($i, 'person_id');
-        $person_ids[$pid] = 1;
-    }
-    $pids = array_keys($person_ids);
-    $pids_str = join(',', $pids);
-
-    $where = '';
-    if ($current_only) {
-        $where = "AND left_house='9999-12-31'";
-    }
-
-    # This is not totally accurate (e.g. minimum entered date may be from a
-    # different house, or similar), but should be good enough.
-    $db = new ParlDB;
-    $q = $db->query("SELECT member.person_id,
-                            title, given_name, family_name, lordofname,
-                            constituency, party,
-                            (SELECT MIN(entered_house) FROM member m WHERE m.person_id=member.person_id) min_entered_house,
-                            left_house, house
-                    FROM member, person_names pn
-                    WHERE member.person_id IN ($pids_str) $where
-                        AND member.person_id = pn.person_id AND pn.type = 'name'
-                        AND pn.start_date <= member.left_house AND member.left_house <= pn.end_date
-                        AND left_house = (SELECT MAX(left_house) FROM member m WHERE m.person_id=member.person_id)
-                    GROUP BY person_id
-                    ORDER BY family_name, lordofname, given_name, person_id");
-
-    return $q;
-}
-
-// Given a search term, find constituencies by name or postcode
-// Returns a list of the array of constituencies, then a boolean saying whether
-// it was a postcode used.
-function search_constituencies_by_query($searchterm) {
-    if (validate_postcode($searchterm)) {
-        // Looks like a postcode - can we find the constituency?
-        $constituency = MySociety\TheyWorkForYou\Utility\Postcode::postcodeToConstituency($searchterm);
-        if ($constituency) {
-            return array( array($constituency), true );
-        }
-    }
-
-    // No luck so far - let's see if they're searching for a constituency.
-    $try = strtolower($searchterm);
-    $query = "select distinct
-            (select name from constituency where cons_id = o.cons_id and main_name) as name
-        from constituency AS o where name like :try
-        and from_date <= date(now()) and date(now()) <= to_date";
-    $db = new ParlDB;
-    $q = $db->query($query, array(':try' => '%' . $try . '%'));
-
-    $constituencies = array();
-    for ($n=0; $n<$q->rows(); $n++) {
-        $constituencies[] = $q->field($n, 'name');
-    }
-
-    return array( $constituencies, false );
-}
